% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapQC.R
\name{REDCapQC}
\alias{REDCapQC}
\title{REDCapQC}
\usage{
REDCapQC(dataset_dir, dataset)
}
\arguments{
\item{dataset_dir}{Directory that holds the location of the dataset and data dictionary}

\item{dataset}{Name of dataset from REDCap}
}
\value{
A data frame with a list of queries.
Variables will include id, form, var, txt (description of query), and report date
}
\description{
Create a QC list from REDCap data dictionary and REDCap all data export
}
